+++ GameOfLife/src/GoL_Engine.h
@@ -15,15 +15,15 @@ public:
 		virtual void update(const Grid& grid) const = 0;
 	};
 
-	GoL_Engine(size_t width_pow2, size_t height_pow2, size_t fill_thresold);
-	GoL_Engine(size_t width_pow2, size_t height_pow2, size_t fill_thresold, size_t max_step);
+	GoL_Engine(int width_pow2, int height_pow2, int fill_thresold);
+	GoL_Engine(int width_pow2, int height_pow2, int fill_thresold, size_t max_step);
 
 	void registerObserver(std::shared_ptr<const IObserver> o);
 
 	void init();
 	void run_serial();
-	void run_omp(size_t nb_thread);
-	void run_std_thread(size_t nb_thread);
+	void run_omp(int nb_thread);
+	void run_std_thread(int nb_thread);
 
 	bool allAlive() const;
 	bool allDead() const;
@@ -33,7 +33,7 @@ private:
 
 	const Grid* do_step();
 	const Grid* do_step_omp();
-	const Grid* do_step_std_thread(size_t nb_thread);
+	const Grid* do_step_std_thread(int nb_thread);
 
 	void kill_lowNeighbours();
 	void live_2to3neighbours();
@@ -46,7 +46,7 @@ private:
 	size_t step_number_;
 	size_t max_step_;
 
-	const size_t fill_thresold_; //% de la grille initialisée à Alive
+	const int fill_thresold_; //% de la grille initialisée à Alive
 
 	std::vector<std::shared_ptr<const IObserver>> observers_;
 };
